The Great Fronted Patch II

+++ collectionsapp/views.py
@@ -37,23 +37,22 @@ def login_user(request):
 def register(request):
     if request.user.is_authenticated:
         return redirect('collections:index')
-
-    form = UserForm(request.POST)
-    if form.is_valid():
-        user = form.save(commit=False)
-        username = form.cleaned_data['username']
-        password = form.cleaned_data['password']
-        user.set_password(password)
-        user.save()
-        user = authenticate(username=username, password=password)
-        if user is not None:
-            if user.is_active:
-                login(request, user)
-                return redirect('collections:index')
-    context = {
-        "form": form,
-    }
-    return render(request, 'collections/register.html', context)
+    if request.method == "POST":
+        form = UserForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            username = form.cleaned_data['username']
+            password = form.cleaned_data['password']
+            user.set_password(password)
+            user.save()
+            user = authenticate(username=username, password=password)
+            if user is not None:
+                if user.is_active:
+                    login(request, user)
+                    return redirect('collections:index')
+    else:
+        form = UserForm()
+    return render(request, 'collections/register.html', {"form": form})
 
 
 def index(request):
@@ -97,10 +96,11 @@ def create_collection(request):
 
 def delete_collection(request, collection_id):
     if request.user.is_authenticated:
-        collection = Collection.objects.get(pk=collection_id)
-        if request.user == collection.user:
-            collection.delete()
-    return redirect("/")
+        if request.method == "POST":
+            collection = Collection.objects.get(pk=collection_id)
+            if request.user == collection.user:
+                collection.delete()
+    return redirect("collections:index")
 
 
 def edit_collection(request, collection_id):
@@ -149,11 +149,13 @@ def create_item(request, collection_id):
 
 
 def delete_item(request, value_id):
+    collection_item = CollectionValue.objects.get(pk=value_id)
+    collection_id = collection_item.collection_id.pk
+
     if request.user.is_authenticated:
-        collection_item = CollectionValue.objects.get(pk=value_id)
-        if request.user == collection_item.collection_id:
-            collection_id = collection_item.collection_id.pk
-            collection_item.delete()
+        if request.user == collection_item.collection_id.user:
+            if request.method == "POST":
+                collection_item.delete()
     return redirect("collections:detail", collection_id=collection_id)
